+++ binding.gyp
@@ -21,7 +21,9 @@
         './src/addon.cpp',
         './src/executor.cpp',
         './src/handle_manager.cpp',
-        './src/rcl_action_bindings.cpp',
+        './src/rcl_action_client_bindings.cpp',
+        './src/rcl_action_goal_bindings.cpp',
+        './src/rcl_action_server_bindings.cpp',
         './src/rcl_bindings.cpp',
         './src/rcl_client_bindings.cpp',
         './src/rcl_context_bindings.cpp',
