+++ src/rcl_action_goal_bindings.h
@@ -1,0 +1,26 @@
+// Copyright (c) 2020 Matt Richard. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef SRC_RCL_ACTION_GOAL_BINDINGS_H_
+#define SRC_RCL_ACTION_GOAL_BINDINGS_H_
+
+#include <napi.h>
+
+namespace rclnodejs {
+
+Napi::Object InitActionGoalBindings(Napi::Env env, Napi::Object exports);
+
+}
+
+#endif  // SRC_RCL_ACTION_GOAL_BINDINGS_H_
